pyproject.toml
[tool.poetry]
name = "transitbuddy-backend"
version = "0.1.0"
description = ""
authors = ["Eric Hanchrow <eric.hanchrow@gmail.com>"]
readme = "README.md"
packages = [{include = "transitbuddy_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
flask-cors = "^3.0.10"
bcrypt = "^4.0.1"
chardet = "^5.1.0"
gtfs-realtime-bindings = "^1.0.0"
python-dotenv = "^1.0.0"
protobuf = "^4.23.2"
protobuf-to-dict = "^0.1.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
